+++ tests/Account_tests.cc
@@ -8,12 +8,12 @@ using namespace account;
 TEST(AccountTests, PaymentTest) {
 	// Arrange
 	
-	std::shared_ptr<Payment>account1(std::make_shared<Payment>("Ilya", 1000.0f));
+	std::shared_ptr<Payment>account1(std::make_shared<Payment>("Ilya", 1000.0));
 		
 	// Act		
 	std::string name = account1->get_name();
-	float balance = account1->get_balance();
-	float accrual = account1->accrual();
+	double balance = account1->get_balance();
+	double accrual = account1->accrual();
 
 	// Assert	
 	EXPECT_EQ(name, "Ilya");
@@ -23,17 +23,17 @@ TEST(AccountTests, PaymentTest) {
 
 TEST(AccountTests, DepositTest) {
 	// Arrange
-	std::shared_ptr<Deposit>account1(std::make_shared<Deposit>("Ilya", 1000.0f, 2.0f));
+	std::shared_ptr<Deposit>account1(std::make_shared<Deposit>("Ilya", 1000, 2.0));
 
 	// Act
 	
-	float percent = account1->get_percent();
+	double percent = account1->get_percent();
 	std::string name = account1->get_name();
-	float balance = account1->get_balance();
-	float accrual = account1->accrual();
+	double balance = account1->get_balance();
+	double accrual = account1->accrual();
 
 	// Assert	
-	EXPECT_EQ(percent, 2.0f);
+	EXPECT_EQ(percent, 2);
 	EXPECT_EQ(name, "Ilya");
 	EXPECT_NEAR(balance, 1000, 0.001);
 	EXPECT_NEAR(accrual, 1020, 0.001);
@@ -41,16 +41,16 @@ TEST(AccountTests, DepositTest) {
 
 TEST(AccountTests, CreditTest) {
 	// Arrange
-	std::shared_ptr<Credit>account1(std::make_shared<Credit>("Ilya", -1000.0f, 2.0f));
+	std::shared_ptr<Credit>account1(std::make_shared<Credit>("Ilya", -1000, 2));
 
 	// Act	
-	float percent = account1->get_percent();
+	double percent = account1->get_percent();
 	std::string name = account1->get_name();
-	float balance = account1->get_balance();
-	float accrual = account1->accrual();
+	double balance = account1->get_balance();
+	double accrual = account1->accrual();
 
 	// Assert	
-	EXPECT_EQ(percent, 2.0f);
+	EXPECT_EQ(percent, 2);
 	EXPECT_EQ(name, "Ilya");
 	EXPECT_NEAR(balance, -1000, 0.001);
 	EXPECT_NEAR(accrual, -1020, 0.001);
